program/requirement.txt
roboflow
numpy
easyocr
ultralytics
mysql-connector-python